Move specific methods from common StringExtensions to the `StringEqualityStrategy` class
